Only set primary for gcc on x86_64.
Due to slightly different compression results from
zlib on arm64 which mean ktxdiff must be used.

Add value of cli_args.primary to command echoed for
re-running a failed subcase so avoid someone going
into a black hole for 2 hours discovering why the
test passes with the echoed command but not when
run via ctests.