Increase tolerance.
These tests fail when run by code compiled with GCC and running on x64_64.
The golden files were regenerated using a binary compiled with GCC 14
on arm64. They pass when run on x64-64 by code compiled with MSVC or Clang.
They also pass when run on arm64 by code compiled with Clang. The differences
are in the zlib result.